registration toggle added

+++ src/controllers/adminControllers.js
@@ -3,32 +3,21 @@ const Project = require('../models/Project')
 const Response = require('../models/Response')
 
 
-const getRegistrationStatus = async (req, res, next) => {
-  try {
-    let admin = await Admin.findOne({});
-    if (!admin) {
-      res.status(500).send({ message: "No admin found!" });
-    }
-    const status = admin.registrationStatus;
-    res.send({ status: status });
-  } catch (err) {
-    res.status(500).send({ message: "Error Fetching Data..." });
-  }
-};
 
 const toggleRegistrationStatus = async (req, res, next) => {
   try {
     let admin = await Admin.findOne({});
     if (!admin) {
-      res.status(500).send({ message: "No admin found!" });
+      return res.status(500).send({ message: "No admin found!" });
     }
-    const {status} = admin.registrationStatus;
-
-    if(status == true){
+    const {registrationStatus} = admin;
+    if(registrationStatus == true){
         admin.registrationStatus = false;
         await admin.save();
-        res.send({message : "Registration closed successfully!"})
+        return res.send({message : "Registration closed successfully!"})
     }
+    admin.registrationStatus = true;
+    await admin.save();
     await Project.deleteMany({})
     await Response.deleteMany({})
     res.send({ message : "Registration started successfully" });
@@ -38,6 +27,5 @@ const toggleRegistrationStatus = async (req, res, next) => {
 };
 
 module.exports = {
-  getRegistrationStatus,
   toggleRegistrationStatus,
 };
